+++ R/ggbarstats.R
@@ -292,41 +292,37 @@ ggbarstats <- function(data,
     # adding significance labels to bars for proportion tests
     if (isTRUE(bar.proptest)) {
       # display grouped proportion test results
-      if (isTRUE(messages)) {
-        # tell the user what these results are
-        proptest_message(rlang::as_name(main), rlang::as_name(condition))
-
-        # print the tibble
-        print(df_labels)
-      }
+      if (isTRUE(messages)) print(df_labels)
 
       # modify plot
-      p <- p + ggplot2::geom_text(
-        data = df_labels,
-        mapping = ggplot2::aes(
-          x = {{ condition }},
-          y = 1.05,
-          label = significance,
-          fill = NULL
-        ),
-        size = 5,
-        na.rm = TRUE
-      )
+      p <- p +
+        ggplot2::geom_text(
+          data = df_labels,
+          mapping = ggplot2::aes(
+            x = {{ condition }},
+            y = 1.05,
+            label = significance,
+            fill = NULL
+          ),
+          size = 5,
+          na.rm = TRUE
+        )
     }
 
     # adding sample size info
     if (isTRUE(sample.size.label)) {
-      p <- p + ggplot2::geom_text(
-        data = df_labels,
-        mapping = ggplot2::aes(
-          x = {{ condition }},
-          y = -0.05,
-          label = N,
-          fill = NULL
-        ),
-        size = 4,
-        na.rm = TRUE
-      )
+      p <- p +
+        ggplot2::geom_text(
+          data = df_labels,
+          mapping = ggplot2::aes(
+            x = {{ condition }},
+            y = -0.05,
+            label = N,
+            fill = NULL
+          ),
+          size = 4,
+          na.rm = TRUE
+        )
     }
 
     # =========================== putting all together ========================