+++ R/reexports.R
@@ -30,3 +30,7 @@ magrittr::`%<>%`
 #' @export
 #' @importFrom magrittr "%$%"
 magrittr::`%$%`
+
+#' @export
+#' @importFrom tibble tibble
+tibble::tibble