+++ tests/testthat/test-switch_statements.R
@@ -7,17 +7,20 @@ testthat::test_that(
   code = {
     testthat::skip_on_cran()
 
-    testthat::expect_error(p.adjust.method.description(NULL))
-    testthat::expect_identical(p.adjust.method.description("none"), "None")
+    testthat::expect_error(p_adjust_text(NULL))
+    testthat::expect_identical(p_adjust_text("none"), "None")
     testthat::expect_identical(
-      p.adjust.method.description("fdr"),
-      p.adjust.method.description("BH")
+      p_adjust_text("fdr"),
+      p_adjust_text("BH")
     )
-    testthat::expect_identical(p.adjust.method.description("xyz"), "Holm")
+    testthat::expect_identical(p_adjust_text("hochberg"), "Hochberg")
+    testthat::expect_identical(p_adjust_text("hommel"), "Hommel")
+    testthat::expect_identical(p_adjust_text("BY"), "Benjamini & Yekutieli")
+    testthat::expect_identical(p_adjust_text("xyz"), "Holm")
   }
 )
 
-# switch for effct size type works ------------------------------------------
+# switch for effect size type works ------------------------------------------
 
 testthat::test_that(
   desc = "switch for effct size type works",